requirements.txt
pandas~=2.3.3
python-dotenv~=1.2.1
requests~=2.32.5
twilio~=9.8.8